[SPARK-46504][PS][TESTS][FOLLOWUPS] Moving move slow tests out of `IndexesTests`
### What changes were proposed in this pull request?
Moving move slow tests out of `IndexesTests`

### Why are the changes needed?
for testing parallelism

### Does this PR introduce _any_ user-facing change?
no

### How was this patch tested?
ci

### Was this patch authored or co-authored using generative AI tooling?
no

Closes #44543 from zhengruifeng/ps_test_idx_base_sort_take.

Authored-by: Ruifeng Zheng <ruifengz@apache.org>
