+++ dev/sparktestsupport/modules.py
@@ -798,6 +798,10 @@ pyspark_pandas_slow = Module(
         "pyspark.pandas.tests.indexes.test_base",
         "pyspark.pandas.tests.indexes.test_conversion",
         "pyspark.pandas.tests.indexes.test_drop",
+        "pyspark.pandas.tests.indexes.test_sort",
+        "pyspark.pandas.tests.indexes.test_symmetric_diff",
+        "pyspark.pandas.tests.indexes.test_take",
+        "pyspark.pandas.tests.indexes.test_unique",
         "pyspark.pandas.tests.indexes.test_asof",
         "pyspark.pandas.tests.indexes.test_astype",
         "pyspark.pandas.tests.indexes.test_delete",
@@ -1090,6 +1094,10 @@ pyspark_pandas_connect_part0 = Module(
         "pyspark.pandas.tests.connect.indexes.test_parity_base",
         "pyspark.pandas.tests.connect.indexes.test_parity_conversion",
         "pyspark.pandas.tests.connect.indexes.test_parity_drop",
+        "pyspark.pandas.tests.connect.indexes.test_parity_sort",
+        "pyspark.pandas.tests.connect.indexes.test_parity_symmetric_diff",
+        "pyspark.pandas.tests.connect.indexes.test_parity_take",
+        "pyspark.pandas.tests.connect.indexes.test_parity_unique",
         "pyspark.pandas.tests.connect.indexes.test_parity_asof",
         "pyspark.pandas.tests.connect.indexes.test_parity_astype",
         "pyspark.pandas.tests.connect.indexes.test_parity_delete",
